+++ Server/src/main/java/server/network/Sender.java
@@ -1,6 +1,6 @@
 package server.network;
 
 public interface Sender {
-    void sendToClient(byte sessionId, byte[] packet);
+    void sendToClient(long sessionId, byte[] packet);
     void sendToRoom(int roomId, byte[] packet);
 }
